refactor: refactor and move user types
