+++ src/lib/user/derive.ts
@@ -1,5 +1,5 @@
+import { AuthSource } from '@/types/user'
 import { v5 as uuidv5, v4 as uuidv4 } from 'uuid'
-import { AuthSource } from './types'
 
 // UUID namespace for GovBot user IDs (generated once with v4)
 const GOVBOT_NAMESPACE = '6ba7b810-9dad-11d1-80b4-00c04fd430c8'
